improved profile user checking and added to all required endpoints

+++ src/main/java/team/bham/web/rest/UserProfileResource.java
@@ -20,7 +20,9 @@ import team.bham.domain.UserProfile;
 import team.bham.repository.UserProfileRepository;
 import team.bham.repository.UserRepository;
 import team.bham.security.SecurityUtils;
+import team.bham.service.UserProfileService;
 import team.bham.web.rest.errors.BadRequestAlertException;
+import team.bham.web.rest.errors.EmailAlreadyUsedException;
 import tech.jhipster.web.util.HeaderUtil;
 import tech.jhipster.web.util.ResponseUtil;
 
@@ -34,15 +36,29 @@ public class UserProfileResource {
 
     private final Logger log = LoggerFactory.getLogger(UserProfileResource.class);
 
+    private static class UserProfileResourceException extends RuntimeException {
+
+        private UserProfileResourceException(String message) {
+            super(message);
+        }
+    }
+
     private static final String ENTITY_NAME = "userProfile";
 
     @Value("${jhipster.clientApp.name}")
     private String applicationName;
 
+    private final UserProfileService userProfileService;
+
     private final UserProfileRepository userProfileRepository;
     private final UserRepository userRepository;
 
-    public UserProfileResource(UserProfileRepository userProfileRepository, UserRepository userRepository) {
+    public UserProfileResource(
+        UserProfileService userProfileService,
+        UserProfileRepository userProfileRepository,
+        UserRepository userRepository
+    ) {
+        this.userProfileService = userProfileService;
         this.userProfileRepository = userProfileRepository;
         this.userRepository = userRepository;
     }
@@ -60,21 +76,13 @@ public class UserProfileResource {
         if (userProfile.getId() != null) {
             throw new BadRequestAlertException("A new userProfile cannot already have an ID", ENTITY_NAME, "idexists");
         }
-
-        Optional<User> userLoggedIn = SecurityUtils.getCurrentUserLogin().flatMap(userRepository::findOneWithAuthoritiesByLogin);
-        if (userLoggedIn.isPresent()) {
-            Long userId = userLoggedIn.get().getId();
-            if (!userProfileRepository.existsById(userId)) {
-                userProfile.setId(userLoggedIn.get().getId());
-                UserProfile result = userProfileRepository.save(userProfile);
-                return ResponseEntity
-                    .created(new URI("/api/user-profiles/" + result.getId()))
-                    .headers(HeaderUtil.createEntityCreationAlert(applicationName, false, ENTITY_NAME, result.getId().toString()))
-                    .body(result);
-            }
-        }
-        log.debug("this should not happen cuz the front end should save us :)");
-        return null;
+        long userId = userProfileService.getUserId();
+        userProfile.setId(userId);
+        UserProfile result = userProfileRepository.save(userProfile);
+        return ResponseEntity
+            .created(new URI("/api/user-profiles/" + result.getId()))
+            .headers(HeaderUtil.createEntityCreationAlert(applicationName, false, ENTITY_NAME, result.getId().toString()))
+            .body(result);
     }
 
     /**
@@ -103,18 +111,17 @@ public class UserProfileResource {
         if (!userProfileRepository.existsById(id)) {
             throw new BadRequestAlertException("Entity not found", ENTITY_NAME, "idnotfound");
         }
-        Optional<User> userLoggedIn = SecurityUtils.getCurrentUserLogin().flatMap(userRepository::findOneWithAuthoritiesByLogin);
-        if (userLoggedIn.isPresent()) {
-            Long userId = userLoggedIn.get().getId();
-            if (Objects.equals(userId, userProfile.getId())) {
-                UserProfile result = userProfileRepository.save(userProfile);
-                return ResponseEntity
-                    .ok()
-                    .headers(HeaderUtil.createEntityUpdateAlert(applicationName, false, ENTITY_NAME, userProfile.getId().toString()))
-                    .body(result);
-            }
+
+        long userId = userProfileService.getUserId();
+        if (!Objects.equals(userId, userProfile.getId())) {
+            throw new UserProfileResource.UserProfileResourceException("Not Authorised"); //replace with 403 response
         }
-        return ResponseEntity.badRequest().body(userProfile);
+
+        UserProfile result = userProfileRepository.save(userProfile);
+        return ResponseEntity
+            .ok()
+            .headers(HeaderUtil.createEntityUpdateAlert(applicationName, false, ENTITY_NAME, userProfile.getId().toString()))
+            .body(result);
     }
 
     /**
@@ -134,6 +141,7 @@ public class UserProfileResource {
         @NotNull @RequestBody UserProfile userProfile
     ) throws URISyntaxException {
         log.debug("REST request to partial update UserProfile partially : {}, {}", id, userProfile);
+
         if (userProfile.getId() == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
@@ -143,6 +151,10 @@ public class UserProfileResource {
 
         if (!userProfileRepository.existsById(id)) {
             throw new BadRequestAlertException("Entity not found", ENTITY_NAME, "idnotfound");
+        }
+        long userId = userProfileService.getUserId();
+        if (!Objects.equals(userId, userProfile.getId())) {
+            throw new UserProfileResource.UserProfileResourceException("Not Authorised"); //replace with 403 response
         }
 
         Optional<UserProfile> result = userProfileRepository
@@ -223,6 +235,10 @@ public class UserProfileResource {
      */
     @DeleteMapping("/user-profiles/{id}")
     public ResponseEntity<Void> deleteUserProfile(@PathVariable Long id) {
+        long userId = userProfileService.getUserId();
+        if (!Objects.equals(userId, id)) {
+            throw new UserProfileResource.UserProfileResourceException("Not Authorised"); //replace with 403 response
+        }
         log.debug("REST request to delete UserProfile : {}", id);
         userProfileRepository.deleteById(id);
         return ResponseEntity
